Add template params, add mapstruct, remove lombok
I prefer the strictness in templates, because it makes its usage a lot clearer - the template tells you what data it needs.

Mapstruct is a consequence of that change, because we will be mapping some database entities to template data and vice versa.

Lombok got removed because there was a little too much code generation going on for my taste. We might regret it, but I prefer having fewer dependencies.

Trimmed down the build.gradle file as well. And h2 database is now only used when running in dev mode, which means it is not included in the production jar either. Added a necessary compiler arg, because the Qute template engine will not compile without it.

Also added some much needed exceptions for the scrapers and crawlers.