Add users, permissions and simple form based auth
This was a little painful to implement, because Quarkus does not support form based auth for reactive apps natively. Quarkus also doesn't support JPA based auth, because that needs JDBC alongside a reactive db connection pool, which cannot be done.

So the solution was to implement some parts of it ourselves, like the UserPassAuth and TrustedAuth classes. The FormBasedAuth class delegates auth requests to the Quarkus built-in FormAuthenticationMechanism, but crucially uses @ActivateRequestContext to make it work in a reactive environment.