Remove reactive as it was a major pain
The reactive libraries caused problems every step of the way. This resulted in tons of workarounds and lots of frustration. Features took forever to implement. Code became very obtuse. Some features were not even possible without major code changes.

Reactive was introduced to make a lightweight application that could handle tons of requests. Most of all, it was added to help parallelize the scraping process. But that turns out unnecessary, because a simple .stream().parallel() fixed that.

With reactive gone, we can also start introducing some libraries that were incompatible before, like elytron for security.