Add search, fix hibernate type issues, fix style
Search functionality uses postgres text search, which we added a GIS index for. Currently, it just searches games by name and does not perform real-time searches. The search page can be expanded in the future by introducing more filters, for things like release date and game designer.

Hibernate was complaining about mismatching data types. The reason is that we use postgres serial4 for auto incrementing primary keys on id columns. They are 4 bytes, whereas hibernate expects id columns to use 8 bytes, so int8 or bigint. This is represented in java by long. But seeing as serial4 is just 4 bytes, we have to use integers, and instruct hibernate that the column is an integer